test: align auth server ownership and traceability
