+++ package.json
@@ -76,7 +76,7 @@
     "test:api-v1:black-box:postgres:crdt": "npm --workspace @ar-eye-hunter/shared-test run bb:api-v1:postgres:crdt",
     "test:api-v1:black-box:postgres:medium-scale": "npm --workspace @ar-eye-hunter/shared-test run bb:api-v1:postgres:medium-scale",
     "test:api-v1:black-box:memory": "npm --workspace @ar-eye-hunter/shared-test run bb:api-v1:memory",
-    "test:repo-governance": "vitest run packages/tests/repo/rallar-authoritative-mutation-guidance-integrity.test.ts packages/tests/repo/rallar-group-state-owner-integrity.test.ts packages/tests/repo/group-state-navigation-map-integrity.test.ts packages/tests/repo/api-v1-group-state-route-lineage-provenance.test.ts packages/tests/repo/client-state-navigation-map-integrity.test.ts packages/tests/repo/client-state-server-lineage-provenance.test.ts packages/tests/repo/client-state-server-persistence-lineage-provenance.test.ts packages/tests/repo/client-state-server-ownership.test.ts packages/tests/repo/client-state-server-ordinary-transaction-lineage-provenance.test.ts packages/tests/repo/client-state-server-test-ownership.test.ts packages/tests/repo/auth-server-lineage-provenance.test.ts packages/tests/repo/auth-server-shell-lineage-provenance.test.ts packages/tests/repo/auth-server-navigation-map-integrity.test.ts packages/tests/repo/auth-server-ownership.test.ts packages/tests/repo/auth-server-test-ownership.test.ts packages/tests/repo/rallar-skill-app-examples-integrity.test.ts packages/tests/repo/rallar-skill-plugin-publication-integrity.test.ts packages/tests/repo/repo-code-style-authority-integrity.test.ts packages/tests/repo/repo-code-style-checker-integrity.test.ts packages/tests/repo/repo-code-style-review-evidence-integrity.test.ts packages/tests/repo/repo-style-changed-check.test.ts packages/tests/repo/repo-style-check.test.ts packages/tests/repo/repo-style-construction-check.test.ts packages/tests/repo/repo-style-construction-edge-cases.test.ts packages/tests/repo/repo-style-layout-rules.test.ts packages/tests/repo/repo-style-structural-lineage-provenance.test.ts packages/tests/rallar-black-box/rallar-testing-skill.test.ts",
+    "test:repo-governance": "vitest run packages/tests/repo/rallar-authoritative-mutation-guidance-integrity.test.ts packages/tests/repo/rallar-group-state-owner-integrity.test.ts packages/tests/repo/group-state-navigation-map-integrity.test.ts packages/tests/repo/api-v1-group-state-route-lineage-provenance.test.ts packages/tests/repo/client-state-navigation-map-integrity.test.ts packages/tests/repo/client-state-server-lineage-provenance.test.ts packages/tests/repo/client-state-server-persistence-lineage-provenance.test.ts packages/tests/repo/client-state-server-ownership.test.ts packages/tests/repo/client-state-server-ordinary-transaction-lineage-provenance.test.ts packages/tests/repo/client-state-server-test-ownership.test.ts packages/tests/repo/auth-server-lineage-provenance.test.ts packages/tests/repo/auth-server-shell-lineage-provenance.test.ts packages/tests/repo/auth-server-navigation-map-integrity.test.ts packages/tests/repo/auth-server-ownership.test.ts packages/tests/repo/auth-server-source-ratchet.test.ts packages/tests/repo/auth-server-test-ownership.test.ts packages/tests/repo/rallar-skill-app-examples-integrity.test.ts packages/tests/repo/rallar-skill-plugin-publication-integrity.test.ts packages/tests/repo/repo-code-style-authority-integrity.test.ts packages/tests/repo/repo-code-style-checker-integrity.test.ts packages/tests/repo/repo-code-style-review-evidence-integrity.test.ts packages/tests/repo/repo-style-changed-check.test.ts packages/tests/repo/repo-style-check.test.ts packages/tests/repo/repo-style-construction-check.test.ts packages/tests/repo/repo-style-construction-edge-cases.test.ts packages/tests/repo/repo-style-layout-rules.test.ts packages/tests/repo/repo-style-structural-lineage-provenance.test.ts packages/tests/rallar-black-box/rallar-testing-skill.test.ts",
     "check:repo-style": "node scripts/repo-style-check.mjs",
     "check:repo-style:construction-details": "node scripts/repo-style-check.mjs --construction-details",
     "check:repo-style:changed": "node scripts/check-changed-repo-style.mjs",
