+++ packages/tests/shared-server/app-inbox-expired-row-replacement.test.ts
@@ -21,7 +21,7 @@ import {
     TestResourceInbox,
     TestResourceInboxResults,
     waitForQueuedEntry,
-} from './app-auth-inbox-test-harness.ts';
+} from './auth/auth-app-inbox-test-runtime.ts';
 import {
     createClientStatePhaseTestDriver,
     failNextClientStateTestOutboxWrite,
