+++ packages/tests/shared-server/auth/auth-command-and-result-codec-regressions.test.ts
@@ -66,176 +66,174 @@ const validDurableResults = [
   },
 ];
 
-describe('auth command and durable-result codec regressions', () => {
-  it('preserves the exact auth mutation rejection error contract', () => {
-    expect(new AuthMutationRejectedError('rejected')).toMatchObject({
-      name: 'AuthMutationRejectedError',
-      message: 'rejected',
-      code: 'auth-mutation-rejected',
-      status: 409,
-    });
-    expect(new AuthMutationRejectedError('unauthorized', 401)).toMatchObject({
-      name: 'AuthMutationRejectedError',
-      message: 'unauthorized',
-      code: 'auth-mutation-rejected',
-      status: 401,
-    });
+it('preserves the exact auth mutation rejection error contract', () => {
+  expect(new AuthMutationRejectedError('rejected')).toMatchObject({
+    name: 'AuthMutationRejectedError',
+    message: 'rejected',
+    code: 'auth-mutation-rejected',
+    status: 409,
   });
-
-  it('strictly decodes every durable auth result variant', () => {
-    for (const result of validDurableResults) {
-      expect(decodeAuthMutationResult(result)).toEqual(result);
-    }
+  expect(new AuthMutationRejectedError('unauthorized', 401)).toMatchObject({
+    name: 'AuthMutationRejectedError',
+    message: 'unauthorized',
+    code: 'auth-mutation-rejected',
+    status: 401,
   });
+});
 
-  it('strictly rejects every malformed durable auth result variant', () => {
-    for (const invalid of [
-      Object.fromEntries(
-        Object.entries(validDurableResults[0]).filter(([key]) => key !== 'requestId'),
-      ),
-      { loggedOut: true },
-      { ...validDurableResults[2], accessToken: 'plaintext-token' },
-      { ...validDurableResults[2], unexpected: true },
-      {
-        kind: 'agent-tickets-issued',
-        tickets: [{ ...validDurableResults[6].tickets![0], ticket: 'x' }],
-      },
-      { kind: 'agent-tickets-issued', tickets: [null] },
-      { ...validDurableResults[3], expiresAtEpochMs: Number.NaN },
-      Object.create({ kind: 'session-issued' }),
-    ]) {
-      expect(() => decodeAuthMutationResult(invalid)).toThrow(TypeError);
-    }
-  });
+it('strictly decodes every durable auth result variant', () => {
+  for (const result of validDurableResults) {
+    expect(decodeAuthMutationResult(result)).toEqual(result);
+  }
+});
 
-  it('captures agent-session lifecycle values before persisted-ticket decoding', () => {
-    let ticketDigestReadCount = 0;
-    let persistedTicketReadStarted = false;
-    const ticket = {
-      agentId: 'agent-1',
-      sessionId: 'agent-session-1',
-      accessTokenDigest: 'agent-access-token-digest',
-      get ticketDigest() {
-        ticketDigestReadCount += 1;
-        if (ticketDigestReadCount === 2) persistedTicketReadStarted = true;
-        return 'agent-ticket-digest';
-      },
+it('strictly rejects every malformed durable auth result variant', () => {
+  for (const invalid of [
+    Object.fromEntries(
+      Object.entries(validDurableResults[0]).filter(([key]) => key !== 'requestId'),
+    ),
+    { loggedOut: true },
+    { ...validDurableResults[2], accessToken: 'plaintext-token' },
+    { ...validDurableResults[2], unexpected: true },
+    {
+      kind: 'agent-tickets-issued',
+      tickets: [{ ...validDurableResults[6].tickets![0], ticket: 'x' }],
+    },
+    { kind: 'agent-tickets-issued', tickets: [null] },
+    { ...validDurableResults[3], expiresAtEpochMs: Number.NaN },
+    Object.create({ kind: 'session-issued' }),
+  ]) {
+    expect(() => decodeAuthMutationResult(invalid)).toThrow(TypeError);
+  }
+});
+
+it('captures agent-session lifecycle values before persisted-ticket decoding', () => {
+  let ticketDigestReadCount = 0;
+  let persistedTicketReadStarted = false;
+  const ticket = {
+    agentId: 'agent-1',
+    sessionId: 'agent-session-1',
+    accessTokenDigest: 'agent-access-token-digest',
+    get ticketDigest() {
+      ticketDigestReadCount += 1;
+      if (ticketDigestReadCount === 2) persistedTicketReadStarted = true;
+      return 'agent-ticket-digest';
+    },
+    clientId: 'client-1',
+    username: 'alice',
+    issuedAtEpochMs: 1_000,
+    get sessionExpiresAtEpochMs() {
+      return persistedTicketReadStarted ? 1_000 : 2_000;
+    },
+    ticketExpiresAtEpochMs: 1_500,
+  };
+  const command = {
+    version: 1,
+    kind: 'issue-agent-tickets',
+    requestId: 'accessor-agent-ticket-command',
+    capturedAtEpochMs: 1_000,
+    authority: {
       clientId: 'client-1',
       username: 'alice',
+      sessionId: 'session-1',
+      accessTokenDigest: 'authority-access-token-digest',
       issuedAtEpochMs: 1_000,
-      get sessionExpiresAtEpochMs() {
-        return persistedTicketReadStarted ? 1_000 : 2_000;
-      },
-      ticketExpiresAtEpochMs: 1_500,
-    };
-    const command = {
-      version: 1,
-      kind: 'issue-agent-tickets',
-      requestId: 'accessor-agent-ticket-command',
-      capturedAtEpochMs: 1_000,
-      authority: {
-        clientId: 'client-1',
-        username: 'alice',
-        sessionId: 'session-1',
-        accessTokenDigest: 'authority-access-token-digest',
-        issuedAtEpochMs: 1_000,
-        expiresAtEpochMs: 3_000,
-      },
-      tickets: [ticket],
-    };
+      expiresAtEpochMs: 3_000,
+    },
+    tickets: [ticket],
+  };
 
-    expect(decodeAuthMutationCommand(command)).toEqual({
-      ...command,
-      tickets: [
-        {
-          ...ticket,
-          sessionExpiresAtEpochMs: 1_000,
-        },
-      ],
-    });
+  expect(decodeAuthMutationCommand(command)).toEqual({
+    ...command,
+    tickets: [
+      {
+        ...ticket,
+        sessionExpiresAtEpochMs: 1_000,
+      },
+    ],
   });
+});
 
-  it('requires an exact registered-user authority on session issuance commands', () => {
-    const command = {
-      version: 1,
-      kind: 'issue-session',
-      requestId: 'session-authority-command',
-      capturedAtEpochMs: 1_000,
+it('requires an exact registered-user authority on session issuance commands', () => {
+  const command = {
+    version: 1,
+    kind: 'issue-session',
+    requestId: 'session-authority-command',
+    capturedAtEpochMs: 1_000,
+    authority: {
+      kind: 'registered-user',
+      clientId: 'client-1',
+      normalizedUsername: 'alice',
+      userRevision: 3,
+    },
+    session: {
+      clientId: 'client-1',
+      username: 'Alice',
+      sessionId: 'session-1',
+      accessTokenDigest: 'digest-1',
+      issuedAtEpochMs: 1_000,
+      expiresAtEpochMs: 2_000,
+    },
+  } as const;
+
+  expect(decodeAuthMutationCommand(command)).toEqual(command);
+  expect(() =>
+    decodeAuthMutationCommand({
+      ...command,
       authority: {
         kind: 'registered-user',
         clientId: 'client-1',
         normalizedUsername: 'alice',
-        userRevision: 3,
       },
-      session: {
-        clientId: 'client-1',
-        username: 'Alice',
-        sessionId: 'session-1',
-        accessTokenDigest: 'digest-1',
-        issuedAtEpochMs: 1_000,
-        expiresAtEpochMs: 2_000,
-      },
-    } as const;
+    }),
+  ).toThrow(TypeError);
+});
 
-    expect(decodeAuthMutationCommand(command)).toEqual(command);
+it('binds session issuance lifecycle to the durable command timestamp', () => {
+  const base = {
+    version: 1,
+    kind: 'issue-session',
+    requestId: 'invalid-session-lifecycle',
+    capturedAtEpochMs: 1_000,
+    authority: {
+      kind: 'static-client',
+      clientId: 'client-1',
+      normalizedUsername: 'alice',
+    },
+    session: {
+      clientId: 'client-1',
+      username: 'alice',
+      sessionId: 'invalid-session',
+      accessTokenDigest: 'digest-1',
+      issuedAtEpochMs: 1_000,
+      expiresAtEpochMs: 2_000,
+    },
+  } as const;
+
+  for (const session of [
+    { ...base.session, issuedAtEpochMs: 999 },
+    { ...base.session, issuedAtEpochMs: 1_001 },
+    { ...base.session, expiresAtEpochMs: 1_000 },
+    { ...base.session, expiresAtEpochMs: 999 },
+  ]) {
     expect(() =>
       decodeAuthMutationCommand({
-        ...command,
-        authority: {
-          kind: 'registered-user',
-          clientId: 'client-1',
-          normalizedUsername: 'alice',
-        },
+        ...base,
+        session,
       }),
-    ).toThrow(TypeError);
-  });
+    ).toThrow(/lifecycle/u);
+  }
+});
 
-  it('binds session issuance lifecycle to the durable command timestamp', () => {
-    const base = {
+it('strictly rejects plaintext or extra auth command fields', () => {
+  expect(() =>
+    decodeAuthMutationCommand({
       version: 1,
-      kind: 'issue-session',
-      requestId: 'invalid-session-lifecycle',
+      kind: 'consume-agent-ticket',
+      requestId: 'consume-1',
       capturedAtEpochMs: 1_000,
-      authority: {
-        kind: 'static-client',
-        clientId: 'client-1',
-        normalizedUsername: 'alice',
-      },
-      session: {
-        clientId: 'client-1',
-        username: 'alice',
-        sessionId: 'invalid-session',
-        accessTokenDigest: 'digest-1',
-        issuedAtEpochMs: 1_000,
-        expiresAtEpochMs: 2_000,
-      },
-    } as const;
-
-    for (const session of [
-      { ...base.session, issuedAtEpochMs: 999 },
-      { ...base.session, issuedAtEpochMs: 1_001 },
-      { ...base.session, expiresAtEpochMs: 1_000 },
-      { ...base.session, expiresAtEpochMs: 999 },
-    ]) {
-      expect(() =>
-        decodeAuthMutationCommand({
-          ...base,
-          session,
-        }),
-      ).toThrow(/lifecycle/u);
-    }
-  });
-
-  it('strictly rejects plaintext or extra auth command fields', () => {
-    expect(() =>
-      decodeAuthMutationCommand({
-        version: 1,
-        kind: 'consume-agent-ticket',
-        requestId: 'consume-1',
-        capturedAtEpochMs: 1_000,
-        ticketDigest: 'digest',
-        ticket: 'plaintext',
-      }),
-    ).toThrow(/fields|plaintext/u);
-  });
+      ticketDigest: 'digest',
+      ticket: 'plaintext',
+    }),
+  ).toThrow(/fields|plaintext/u);
 });
