+++ packages/tests/shared-server/mutation-route-owner-provenance.test.ts
@@ -8,132 +8,131 @@ import {
   validateMutationRouteInventory,
 } from './mutation-routing-inventory.ts';
 
+// Retain permanently as cross-domain semantic route-provenance evidence.
 const CAPABILITY_FIXTURES =
   'packages/tests/shared-server/fixtures/mutation-boundary-capability-receivers';
 
-describe('Mutation route owner provenance contracts', { timeout: 30_000 }, () => {
-  it('follows mutable capability provenance through production receiver shapes', () => {
-    for (const name of [
-      'parameter.ts',
-      'bracket.ts',
-      'constructor.ts',
-      'declared-property.ts',
-      'destructured.ts',
-    ]) {
-      const root = `${CAPABILITY_FIXTURES}/${name}`;
-      expect(findMutationBoundaryViolationsFromRoots([root]), root).toEqual([
-        expect.objectContaining({
-          filePath: root,
-          directMutatorCalls: ['ClientStateRepository.insertPrincipal'],
-        }),
-      ]);
-    }
-    expect(
-      findMutationBoundaryViolationsFromRoots([`${CAPABILITY_FIXTURES}/read-only.ts`]),
-    ).toEqual([]);
-  });
+it('follows mutable capability provenance through production receiver shapes', () => {
+  for (const name of [
+    'parameter.ts',
+    'bracket.ts',
+    'constructor.ts',
+    'declared-property.ts',
+    'destructured.ts',
+  ]) {
+    const root = `${CAPABILITY_FIXTURES}/${name}`;
+    expect(findMutationBoundaryViolationsFromRoots([root]), root).toEqual([
+      expect.objectContaining({
+        filePath: root,
+        directMutatorCalls: ['ClientStateRepository.insertPrincipal'],
+      }),
+    ]);
+  }
+  expect(findMutationBoundaryViolationsFromRoots([`${CAPABILITY_FIXTURES}/read-only.ts`])).toEqual(
+    [],
+  );
+});
 
-  it('rejects a dead correct type after an HTTP handler is given the wrong handoff type', () => {
-    const item = requireEntry(AppInboxType.CLIENT_PRINCIPAL_UPSERT, 'HTTP');
-    const source = readFileSync(item.sourcePath, 'utf8');
-    const mutated = source
-      .replace(
-        'type: AppInboxType.CLIENT_PRINCIPAL_UPSERT,',
-        'type: AppInboxType.CLIENT_INSTANCE_UPSERT,',
-      )
-      .replace(
-        'const request = withActor(requestBody, authSession);',
-        `const request = withActor(requestBody, authSession);
+it('rejects a dead correct type after an HTTP handler is given the wrong handoff type', () => {
+  const item = requireEntry(AppInboxType.CLIENT_PRINCIPAL_UPSERT, 'HTTP');
+  const source = readFileSync(item.sourcePath, 'utf8');
+  const mutated = source
+    .replace(
+      'type: AppInboxType.CLIENT_PRINCIPAL_UPSERT,',
+      'type: AppInboxType.CLIENT_INSTANCE_UPSERT,',
+    )
+    .replace(
+      'const request = withActor(requestBody, authSession);',
+      `const request = withActor(requestBody, authSession);
         const deadCorrectType = (): void => {
           void AppInboxType.CLIENT_PRINCIPAL_UPSERT;
           void deps.processClientAppInbox;
         };
         void deadCorrectType;`,
-      );
-
-    expect(validateWithOverride(item.sourcePath, mutated)).toEqual(
-      expect.arrayContaining([expect.stringContaining('registered handler is not connected')]),
     );
-  });
 
-  it('rejects a rerouted websocket callback even when dead markers remain', () => {
-    const item = requireEntry(AppInboxType.RTC_RTT_SUBMIT, 'WS_INBOX');
-    const source = readFileSync(item.sourcePath, 'utf8');
-    const mutated =
-      source.replace('await enqueueRtcRttMutation({', 'await Promise.resolve({') +
-      `
+  expect(validateWithOverride(item.sourcePath, mutated)).toEqual(
+    expect.arrayContaining([expect.stringContaining('registered handler is not connected')]),
+  );
+});
+
+it('rejects a rerouted websocket callback even when dead markers remain', () => {
+  const item = requireEntry(AppInboxType.RTC_RTT_SUBMIT, 'WS_INBOX');
+  const source = readFileSync(item.sourcePath, 'utf8');
+  const mutated =
+    source.replace('await enqueueRtcRttMutation({', 'await Promise.resolve({') +
+    `
 function deadWsHandoff(enqueueRtcRttMutation: (...args: never[]) => unknown): void {
   void enqueueRtcRttMutation;
   void AppInboxType.RTC_RTT_SUBMIT;
 }
 `;
 
-    expect(validateWithOverride(item.sourcePath, mutated)).toEqual(
-      expect.arrayContaining([expect.stringContaining('registered handler is not connected')]),
-    );
-  });
+  expect(validateWithOverride(item.sourcePath, mutated)).toEqual(
+    expect.arrayContaining([expect.stringContaining('registered handler is not connected')]),
+  );
+});
 
-  it('rejects a lifecycle type dispatched to the wrong owner with a dead correct call', () => {
-    const item = requireEntry(AppInboxType.CLIENT_AUTHORISED_WS_DISCONNECT, 'WS_LIFECYCLE');
-    const source = readFileSync(item.dispatchSourcePath, 'utf8');
-    const live = 'await this.handler.processAuthorisedWsDisconnect(input, context)';
-    const mutated =
-      source.replace(live, 'await this.handler.processAuthorisedWsConnect(input, context)') +
-      `
+it('rejects a lifecycle type dispatched to the wrong owner with a dead correct call', () => {
+  const item = requireEntry(AppInboxType.CLIENT_AUTHORISED_WS_DISCONNECT, 'WS_LIFECYCLE');
+  const source = readFileSync(item.dispatchSourcePath, 'utf8');
+  const live = 'await this.handler.processAuthorisedWsDisconnect(input, context)';
+  const mutated =
+    source.replace(live, 'await this.handler.processAuthorisedWsConnect(input, context)') +
+    `
 class DeadLifecycleOwner {
   process(): void { void this.handler.processAuthorisedWsDisconnect; }
   private readonly handler = { processAuthorisedWsDisconnect(): void {} };
 }
 `;
-    expect(source).toContain(live);
+  expect(source).toContain(live);
 
-    expect(validateWithOverride(item.dispatchSourcePath, mutated)).toEqual(
-      expect.arrayContaining([expect.stringContaining('owner dispatch is not connected')]),
-    );
-  });
+  expect(validateWithOverride(item.dispatchSourcePath, mutated)).toEqual(
+    expect.arrayContaining([expect.stringContaining('owner dispatch is not connected')]),
+  );
+});
 
-  it('rejects a cross-file admin handoff with the wrong type and dead correct evidence', () => {
-    const item = MUTATION_ROUTE_INVENTORY.find((entry) =>
-      entry.entrypoint.includes('/topology/recompute'),
-    );
-    if (!item) throw new Error('Admin topology route is absent');
-    const source = readFileSync(item.enqueueSourcePath, 'utf8');
-    const mutated =
-      source.replace(
-        'type: AppInboxType.TOPOLOGY_RECONFIGURE,',
-        'type: AppInboxType.TOPOLOGY_CONFIG_PUT,',
-      ) + '\nfunction deadAdminType(): void { void AppInboxType.TOPOLOGY_RECONFIGURE; }\n';
+it('rejects a cross-file admin handoff with the wrong type and dead correct evidence', () => {
+  const item = MUTATION_ROUTE_INVENTORY.find((entry) =>
+    entry.entrypoint.includes('/topology/recompute'),
+  );
+  if (!item) throw new Error('Admin topology route is absent');
+  const source = readFileSync(item.enqueueSourcePath, 'utf8');
+  const mutated =
+    source.replace(
+      'type: AppInboxType.TOPOLOGY_RECONFIGURE,',
+      'type: AppInboxType.TOPOLOGY_CONFIG_PUT,',
+    ) + '\nfunction deadAdminType(): void { void AppInboxType.TOPOLOGY_RECONFIGURE; }\n';
 
-    expect(validateWithOverride(item.enqueueSourcePath, mutated)).toEqual(
-      expect.arrayContaining([expect.stringContaining('registered handler is not connected')]),
-    );
-  });
+  expect(validateWithOverride(item.enqueueSourcePath, mutated)).toEqual(
+    expect.arrayContaining([expect.stringContaining('registered handler is not connected')]),
+  );
+});
 
-  it('rejects a cross-file auth handoff discriminator and owner reroute', () => {
-    const item = requireEntry(AppInboxType.AUTH_USER_REGISTER, 'HTTP');
-    const typeOwnerSource = readFileSync(item.typeOwnerSourcePath, 'utf8');
-    const dispatchSource = readFileSync(item.dispatchSourcePath, 'utf8');
-    const wrongType =
-      typeOwnerSource.replace("kind: 'register-user',", "kind: 'issue-session',") +
-      "\nfunction deadAuthType(): void { void 'register-user'; }\n";
-    const wrongOwner =
-      dispatchSource.replace(
-        'await this.authInboxHandler.processAuthMutation(command, context)',
-        'await Promise.resolve(command)',
-      ) +
-      '\nclass DeadAuthOwner { process(): void { void this.authInboxHandler.processAuthMutation; } private readonly authInboxHandler = { processAuthMutation(): void {} }; }\n';
-    expect(typeOwnerSource).toContain("kind: 'register-user',");
-    expect(dispatchSource).toContain(
+it('rejects a cross-file auth handoff discriminator and owner reroute', () => {
+  const item = requireEntry(AppInboxType.AUTH_USER_REGISTER, 'HTTP');
+  const typeOwnerSource = readFileSync(item.typeOwnerSourcePath, 'utf8');
+  const dispatchSource = readFileSync(item.dispatchSourcePath, 'utf8');
+  const wrongType =
+    typeOwnerSource.replace("kind: 'register-user',", "kind: 'issue-session',") +
+    "\nfunction deadAuthType(): void { void 'register-user'; }\n";
+  const wrongOwner =
+    dispatchSource.replace(
       'await this.authInboxHandler.processAuthMutation(command, context)',
-    );
+      'await Promise.resolve(command)',
+    ) +
+    '\nclass DeadAuthOwner { process(): void { void this.authInboxHandler.processAuthMutation; } private readonly authInboxHandler = { processAuthMutation(): void {} }; }\n';
+  expect(typeOwnerSource).toContain("kind: 'register-user',");
+  expect(dispatchSource).toContain(
+    'await this.authInboxHandler.processAuthMutation(command, context)',
+  );
 
-    expect(validateWithOverride(item.typeOwnerSourcePath, wrongType)).toEqual(
-      expect.arrayContaining([expect.stringContaining('registered handler is not connected')]),
-    );
-    expect(validateWithOverride(item.dispatchSourcePath, wrongOwner)).toEqual(
-      expect.arrayContaining([expect.stringContaining('owner dispatch is not connected')]),
-    );
-  });
+  expect(validateWithOverride(item.typeOwnerSourcePath, wrongType)).toEqual(
+    expect.arrayContaining([expect.stringContaining('registered handler is not connected')]),
+  );
+  expect(validateWithOverride(item.dispatchSourcePath, wrongOwner)).toEqual(
+    expect.arrayContaining([expect.stringContaining('owner dispatch is not connected')]),
+  );
 });
 
 function requireEntry(type: AppInboxType, transport: string) {
